fix: correct final bugs
