+++ src/utils.js
@@ -1,3 +1,5 @@
+import { brownCards, blueCards, greenCards } from '../data/mythicCards';
+
 ////////// --- FILL DOTS OF EVERY STAGE --- //////////
 function fillDots(ancient, stages) {
 
@@ -102,4 +104,97 @@ function cardsCount(ancient) {
   }
   //////////////////////////////////////////////////////////////////////////////
 
-  export {fillDots, stageCards, cardsCount};
+  ////////// --- RETURN SET OF CARDS DUE TO DIFICULTY RESTRICTIONS --- //////////
+  function modifiedSetOfCards(diff, numberOfCards) {
+    switch (diff) {
+      case 'very-easy': {
+        let greenEasy = greenCards.filter(val => val.difficulty == 'easy');
+        let greenNormal = greenCards.filter(val => val.difficulty == 'normal');
+        if (numberOfCards.greenCards > greenEasy.length) {
+          greenNormal = _.shuffle(greenNormal);
+          greenNormal = greenNormal.slice(0, numberOfCards.greenCards - greenEasy.length);
+          greenEasy.push(...greenNormal);
+        }
+  
+        let brownEasy = brownCards.filter(val => val.difficulty == 'easy');
+        let brownNormal = brownCards.filter(val => val.difficulty == 'normal');
+        console.log(numberOfCards.brownCards);
+        console.log(brownEasy.length);
+        if (numberOfCards.brownCards > brownEasy.length) {
+          brownNormal = _.shuffle(brownNormal);
+  
+          brownNormal = brownNormal.slice(0, numberOfCards.brownCards - brownEasy.length);
+          brownEasy.push(...brownNormal);
+        }
+  
+        let blueEasy = blueCards.filter(val => val.difficulty == 'easy');
+        let blueNormal = blueCards.filter(val => val.difficulty == 'normal');
+        if (numberOfCards.blueCards > blueEasy.length) {
+          blueNormal = _.shuffle(blueNormal);
+          blueNormal = blueNormal.slice(0, numberOfCards.blueCards - blueEasy.length);
+          blueEasy.push(...blueNormal);
+        }
+          return {
+          greenCards: greenEasy,
+          brownCards: brownEasy,
+          blueCards: blueEasy,
+        }
+      }
+      case 'easy':
+        return {
+          greenCards: greenCards.filter(val => val.difficulty !== 'hard'),
+          brownCards: brownCards.filter(val => val.difficulty !== 'hard'),
+          blueCards: blueCards.filter(val => val.difficulty !== 'hard'),
+        }
+      case 'normal':
+        return {
+          greenCards: greenCards,
+          brownCards: brownCards,
+          blueCards: blueCards
+        }
+      case 'hard':
+        return {
+          greenCards: greenCards.filter(val => val.difficulty !== 'easy'),
+          brownCards: brownCards.filter(val => val.difficulty !== 'easy'),
+          blueCards: blueCards.filter(val => val.difficulty !== 'easy'),
+        }
+      case 'very-hard': {
+        let greenHard = greenCards.filter(val => val.difficulty == 'hard');
+        let greenNormal = greenCards.filter(val => val.difficulty == 'normal');
+        if (numberOfCards.greenCards > greenHard.length) {
+          greenNormal = _.shuffle(greenNormal);
+          greenNormal = greenNormal.slice(0, numberOfCards.greenCards - greenHard.length);
+          greenHard.push(...greenNormal);
+        }
+  
+        let brownHard = brownCards.filter(val => val.difficulty == 'hard');
+        let brownNormal = brownCards.filter(val => val.difficulty == 'normal');
+        if (numberOfCards.brownCards > brownHard.length) {
+          brownNormal = _.shuffle(brownNormal);
+          brownNormal = brownNormal.slice(0, numberOfCards.brownCards - brownHard.length);
+          brownHard.push(...brownNormal);
+        }
+  
+        let blueHard = blueCards.filter(val => val.difficulty == 'hard');
+        let blueNormal = blueCards.filter(val => val.difficulty == 'normal');
+        if (numberOfCards.blueCards > blueHard.length) {
+          blueNormal = _.shuffle(blueNormal);
+          blueNormal = blueNormal.slice(0, numberOfCards.blueCards - blueHard.length);
+          blueHard.push(...blueNormal);
+        }
+          return {
+          greenCards: greenHard,
+          brownCards: brownHard,
+          blueCards: blueHard,
+        }
+      }
+      default:
+        return {
+          greenCards: greenCards,
+          brownCards: brownCards,
+          blueCards: blueCards
+        }
+    }
+  }
+
+  export {fillDots, stageCards, cardsCount, modifiedSetOfCards};
